backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pydantic
httpx
aiosqlite
apscheduler
feedparser
sqlalchemy
asyncpg
